Change deferred local $refs' resolution strategy.
When a SchemaDocument is walked entirely/recursively, it is garanteed
that at the end of the parsing all the local $refs can be pointed to
Schemas already created (unless they are dangling and can be ignored).
So the current strategy to prevent infinite recursion within recursive
Schemas is to defer the resolution of local $refs like this, saving them
temporarily in the SchemaRef_ stack.

But this does not work with SchemaDocuments created from a Pointer (some
specific part of the source document only), because the parsing solely
follows the $refs and thus they must be resolved immediately.

So instead this change adds each Schema and its $refs to SchemaMap_ as
soon it is created (in the Schema constructor itself), and since schemas
are reused from the map if they already exist it also prevents infinite
recursion.

SchemaRef_ is now used to store the $ref(s)/Pointer(s) followed to reach
the actual Schema, all of which are added likewise to the map so that
they also resolve directly to it. This is done with AddSchemaRefs().

The SchemaPointer test is adapted to contain recursive schemas.